+++ vcd/resource_vcd_vapp_org_network.go
@@ -121,7 +121,7 @@ func resourceVappOrgNetworkCreate(d *schema.ResourceData, meta interface{}) erro
 	if err != nil {
 		return fmt.Errorf("unable to get network ID from HREF: %s", err)
 	}
-	d.SetId(networkId)
+	d.SetId(normalizeId("urn:vcloud:network:", networkId))
 
 	return resourceVappOrgNetworkRead(d, meta)
 }
@@ -174,7 +174,7 @@ func genericVappOrgNetworkRead(d *schema.ResourceData, meta interface{}, origin 
 
 	// needs to set for datasource
 	if d.Id() == "" {
-		d.SetId(networkId)
+		d.SetId(normalizeId("urn:vcloud:network:", networkId))
 	}
 
 	_ = d.Set("retain_ip_mac_enabled", *vAppNetwork.Configuration.RetainNetInfoAcrossDeployments)
@@ -308,7 +308,7 @@ func resourceVcdVappOrgNetworkImport(d *schema.ResourceData, meta interface{}) (
 		return nil, fmt.Errorf("unable to get network ID from HREF: %s", err)
 	}
 
-	d.SetId(networkId)
+	d.SetId(normalizeId("urn:vcloud:network:", networkId))
 
 	if vcdClient.Org != orgName {
 		_ = d.Set("org", orgName)