+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 2.9.0 (Unreleased)
+FEATURES:
+
+* **New Resource:** `resource/vcd_vapp_firewall_rules` vApp network firewall rules [GH-511]
 
 FEATURES: